+++ NEWS.md
@@ -1,6 +1,6 @@
 # meditate 0.1.2.9000
 
-- ...
+- Add `PlotMandala`, draws a colored mandala.
 
 # meditate 0.1.2