+++ NEWS.md
@@ -6,4 +6,4 @@
 
 # meditate 0.1.2
 
-- First version of **meditate**
+- Initial CRAN release.