+++ NEWS.md
@@ -1,6 +1,6 @@
 # meditate 0.1.2.9000
 
-- In `Meditate`, add new arguments `mandala` and `user_ends`.
+- In `Meditate`, add new arguments `mandala` and `user_stops`.
 
 - In `ReadSessions`, fix bug with current streak not reporting correctly.