+++ NEWS.md
@@ -1,4 +1,4 @@
-# meditate 0.1.2.9000
+# meditate 0.1.3
 
 - In `Meditate`, add new arguments `mandala` and `user_stops`.